+++ vpmobil/__init__.py
@@ -1,5 +1,5 @@
 """
-API wrapper module for interacting with vpmobil substitution plans
+Python API wrapper library for evaluating VpMobil substitution plans
 
 [[Repository]](https://github.com/annhilati/vpmobil-py)
 [[Documentation]](https://annhilati.github.io/vpmobil-py/)
@@ -14,14 +14,12 @@ from vpmobil.models import (
     )
 from vpmobil.parser import Parser
 
-_symbols = [
-    VertretungsplanZugang,
-    Vertretungsplan,
-    Klasse, Lehrer, Raum,
-    Stunde, Kurs, Aufsicht, Klausur,
-    VpMobilPyError, Unauthorized, ResourceNotFound,
-    Standardpfade, Parser
-]
-_constants = []
 
-__all__ = [obj.__name__ for obj in _symbols] + _constants
+__all__ = [
+    "VertretungsplanZugang",
+    "Vertretungsplan",
+    "Klasse", "Lehrer", "Raum",
+    "Stunde", "Kurs", "Aufsicht", "Klausur",
+    "VpMobilPyError", "Unauthorized", "ResourceNotFound",
+    "Standardpfade", "Parser"
+]