+++ vpmobil/extensions/parser_presets.py
@@ -4,7 +4,7 @@ from vpmobil.parser import Parser
 StBennoGymnasium = Parser(
     AUFZÄHLUNGS_TRENNZEICHEN  = " ",
     KLASSENBEZEICHNER_PATTERN = r"(?P<stufe>[1-9][0-9]?)(?P<suffix>[a-z])",
-    STUNDE_HERVERLEGT_PATTERN = r"verlegt von St\.(?P<periode>\d+);"
+    STUNDE_VERLEGT_VON_PATTERN = r"verlegt von St\.(?P<periode>\d+);"
 )
 
 TolkewitzCampus = Parser(